Move `connection_manager` to `provider`
