+++ docs/_source/docs/hooks.rst
@@ -146,7 +146,7 @@ custom_hook.py
             argument defined in the Sceptre config file (see below)
         stack: sceptre.provider.stack.Stack
              The associated stack of the hook.
-        connection_manager: sceptre.connection_manager.ConnectionManager
+        connection_manager: sceptre.provider.connection_manager.ConnectionManager
             Boto3 Connection Manager - can be used to call boto3 api.
 
         """
