+++ sceptre/provider/connection_manager.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 """
-sceptre.connection_manager
+sceptre.provider.connection_manager
 
 This module implements a ConnectionManager class, which simplifies and manages
 Boto3 calls.
@@ -16,7 +16,7 @@ import boto3
 from os import environ
 from botocore.exceptions import ClientError
 
-from .helpers import mask_key
+from sceptre.helpers import mask_key
 from sceptre.exceptions import InvalidAWSCredentialsError, RetryLimitExceededError
 
 
@@ -90,7 +90,7 @@ class ConnectionManager(object):
 
     def __repr__(self):
         return (
-            "sceptre.connection_manager.ConnectionManager(region='{0}', "
+            "sceptre.provider.connection_manager.ConnectionManager(region='{0}', "
             "profile='{1}', stack_name='{2}')".format(
                 self.region, self.profile, self.stack_name
             )
